rename: Loading, Error 공통컴포넌트로 분리
